backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Ranjith",email = "ranjithrd03@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlmodel (>=0.0.24,<0.0.25)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "minio (>=7.2.16,<8.0.0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "replicate (>=1.0.7,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
